+++ Sources/Models/ALKConfiguration.swift
@@ -91,6 +91,9 @@ public struct ALKConfiguration {
     
     /// For the Form Submit Button, hide post CTA is false by default.
     public var hidePostFormSubmit = false
+    
+    /// For the disabling of form post-submit, the form will not be editable after submission. By default, it's set to false.
+    public var disableFormPostSubmit = false
 
     /// Status bar style. It will be used in all view controllers.
     /// Default value is lightContent.